% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_exploRGO_fct_exploRGO.R
\name{add_selected_GO_results}
\alias{add_selected_GO_results}
\title{add_selected_GO_results}
\usage{
add_selected_GO_results(DGE_data, go_results_selected)
}
\arguments{
\item{DGE_data}{results of \code{\link{filter_DGE_results}}}

\item{go_results_selected}{results of \code{\link{add_selected_GO_results}}}
}
\value{
a tibble
}
\description{
combined the results of \code{\link{filter_DGE_results}} with
 \code{\link{add_selected_GO_results}} to allow highlighting of genes
  in selected pathways
}
