% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_exploRGO_fct_exploRGO.R
\name{bar_chart}
\alias{bar_chart}
\title{bar_chart}
\usage{
bar_chart(
  label,
  width = "100\%",
  height = "14px",
  fill = "#00bfc4",
  background = NULL
)
}
\arguments{
\item{label}{label text}

\item{width}{width of the bar}

\item{height}{height of the bar}

\item{fill}{colour of the bar}

\item{background}{background colour of the cell}
}
\value{
html
}
\description{
bar charts for reactable 'sparklin' plots, fixed color
}
