% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_exploRGO_fct_exploRGO.R
\name{collapse_selected_GO_results_to_label}
\alias{collapse_selected_GO_results_to_label}
\title{collapse_selected_GO_results_to_label}
\usage{
collapse_selected_GO_results_to_label(DGE_with_selected_GO_results)
}
\arguments{
\item{DGE_with_selected_GO_results}{results of \code{\link{add_selected_GO_results}}}
}
\value{
a tibble
}
\description{
collapse_selected_GO_results_to_label
}
\examples{
\dontrun{
selected_GO_results_labelled <- collapse_selected_GO_results_to_label(
	DGE_with_selected_GO_results
)
}
}
