% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_exploRGO_fct_exploRGO.R
\name{counts_selection}
\alias{counts_selection}
\title{counts_selection}
\usage{
counts_selection(counts, comparisons, genes, mean_of_repeats = TRUE)
}
\arguments{
\item{counts}{count matrix as tibble with gene_id, SYMBOL columns}

\item{comparisons}{the comparisons between conditions for which to get the counts}

\item{genes}{the genes for which to get the counts}

\item{mean_of_repeats}{if TRUE (default) takes the mean of repeats, 
if FALSE shows results for all conditions.}
}
\value{
matrix
}
\description{
counts_selection
}
