% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_exploRGO_fct_exploRGO.R
\name{filter_DGE_results}
\alias{filter_DGE_results}
\title{filter_DGE_results}
\usage{
filter_DGE_results(
  dge_data,
  comparisons,
  genes,
  p_value = 0.05,
  lfc_down = -1,
  lfc_up = 1
)
}
\arguments{
\item{dge_data}{differentially expressed gene results from our RNA-seq pipeline}

\item{comparisons}{the comparisons between conditions to include}

\item{genes}{the genes to include}

\item{p_value}{the p-value cut-off}

\item{lfc_down}{the lower log2 fold change threshold}

\item{lfc_up}{the upper log2 fold change threshold}
}
\value{
a tibble
}
\description{
filtering fo the differentiall epxressed gene results
}
