% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_exploRGO_fct_exploRGO.R
\name{get_conditions}
\alias{get_conditions}
\title{get_conditions}
\usage{
get_conditions(comparisons)
}
\arguments{
\item{comparisons}{a list of comparison strings}
}
\value{
a character vector of experimenta conditions
}
\description{
parses comparisons to get conditions
}
