% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_exploRGO_fct_exploRGO.R
\name{get_go_results_selected}
\alias{get_go_results_selected}
\title{get_go_results_selected}
\usage{
get_go_results_selected(go_results_filtered, selected_lines)
}
\arguments{
\item{go_results_filtered}{the GO ORA results after filtering with
\code{\link{go_ora_filters}}}

\item{selected_lines}{the lines from the ORA result table which have been
selected by the user}
}
\value{
a tibble
}
\description{
get_go_results_selected
}
\examples{
\dontrun{
go_results_selected <- get_go_results_selected(
	go_results_filtered, c(1,4,5)
)
}
}
