% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_exploRGO_fct_exploRGO.R
\name{get_go_results_selected_genes}
\alias{get_go_results_selected_genes}
\title{get_go_results_selected_genes}
\usage{
get_go_results_selected_genes(go_results_selected)
}
\arguments{
\item{go_results_selected}{gets just the gene names from the output of
\code{\link{get_go_results_selected}}}
}
\value{
a character vector of gene symbols
}
\description{
get_go_results_selected_genes
}
\examples{
\dontrun{
go_results_selected_genes <- get_go_results_selected_genes(
	 go_results_selected
)
}
}
