% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_exploRGO_fct_exploRGO.R
\name{go_ora_filters}
\alias{go_ora_filters}
\title{go_ora_filters}
\usage{
go_ora_filters(data, comparisons, ontologies, directions, p_value)
}
\arguments{
\item{data}{Gene Ontology (GO) terms over representation analysis (ORA) results derived from clusterProfiler results}

\item{comparisons}{the conditions to display,
from those being compared in the differential gene expression (DGE) analysis on which this ORA is based}

\item{ontologies}{the ontology(s) to display (BP, MF, CC)}

\item{directions}{the direction(s) of change in gene expression for which the ORA was performed (Up, Down, Either)}

\item{p_value}{the p-value cut off for the ORA analysis}
}
\value{
tibble
}
\description{
go_ora_filters
}
