% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_exploRGO_fct_exploRGO.R
\name{heatmap_plotter}
\alias{heatmap_plotter}
\title{heatmap_plotter}
\usage{
heatmap_plotter(mat)
}
\arguments{
\item{mat}{a gene expression matrix}
}
\value{
a ComplexHeatmap Heatmap
}
\description{
heatmap_plotter
}
\examples{
\dontrun{
heatmap_plotter(selected_vst_mat)
}

}
