% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_exploRGO_fct_exploRGO.R
\name{scale_check_set_default_on_fail}
\alias{scale_check_set_default_on_fail}
\title{scale_check_set_default_on_fail}
\usage{
scale_check_set_default_on_fail(vec, default)
}
\arguments{
\item{vec}{a numeric vector}

\item{default}{the default range to use in the event the vector fails checks}
}
\value{
a numeric vector
}
\description{
Checks to see if a vector with have properties that will be a problem for
using it to create a colour scale such as containing infinite values or
having a length of less than 3
allows the user to set a defaul scale if checks fail
}
