% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_exploRGO_fct_exploRGO.R
\name{volcano_plotter}
\alias{volcano_plotter}
\title{volcano_plotter}
\usage{
volcano_plotter(DGE_data)
}
\arguments{
\item{DGE_data}{differential gene expression datr from which to make a plot}
}
\value{
a ggplot2 object
}
\description{
volcano_plotter
}
\examples{
\dontrun{
volcano_plotter(selected_GO_results_labelled)
}
}
